Use the artifact versions specified in the root module when resolving
Before this, it was effectively random which artifact would be used
in the dependency resolution. This change means that the root module
can force specific versions to be used.

The drawback is that because we now set specific versions, if an
`install` tag has the same dependency at different versions the
resolver will now no longer warn of duplicate entries, though the
new code will emit a similar warning, so no information has been
lost.